requirements.txt
Django==2.1.5
django-debug-toolbar==1.11
django-extensions==2.1.4
selenium==2.32.0
simplejson==3.1.2
six==1.2.0
termcolor==1.1.0
wsgiref==0.1.2
coverage==3.6
Sphinx==1.1.3
